Setup testing and lint environment
This commit adds the basic pieces necessary to actually run the unit
tests in CI and setup linting checks. Moving forward all PRs will be
required to pass both in CI on all supported environments prior to
merging. Right now only a single pro-forma test runs since the tests
included in the initial implementation of the provider require backend
access, which can not be a requirement for CI. Those tests are now
skipped unless credentials are present.