Add CI and Test setup (#3)
* Setup testing and lint environment

This commit adds the basic pieces necessary to actually run the unit
tests in CI and setup linting checks. Moving forward all PRs will be
required to pass both in CI on all supported environments prior to
merging. Right now only a single pro-forma test runs since the tests
included in the initial implementation of the provider require backend
access, which can not be a requirement for CI. Those tests are now
skipped unless credentials are present.

* Fix all lint issues

This commit goes through all the lint failures in the repo and fixes
them. After this point we should start enforcing lint on all incoming
PRs.

* Relax websockets requirement slightly

Qiskit still supports python 3.5 (until the upstream Python EoL date in
September), however the minimum version of the websockets library
>=8.0.0 does not include support for python 3.5. This commit relaxes the
minimum version to enable using 0.7.0 when running with python 3.5.