Update and fix python package and versioning
This commit goes through and fixes several aspects related to the
packaging around the honeywell provider. It mainly fixes small
oversights that were resulting in a package that would work but was
either including extra files which were unecsessary or missing metadata.
At the same time the versioning scheme for the project is updated to be
consistent with the rest of Qiskit. A __version__ attribute is added to
__init__ module for the honeywell namespace too.