Remove marshmallow import and fix exception catch
Right now marshmallow is not part of the honeywell provider's
requirements list, however it's being directly imported in one place and
only used for catching an exception (actually a subclass of an
exception). Terra is moving away from using marshmallow so this commit
removes the sole marshmallow import and updates the exception catch to
use the base qiskit exception class which will still work and not be
removed.