Squashed commit including basis update and proxy support. (#11)
The set of basis gates was insufficient to allow submission to our system. By adding additional allowed basis gates we were able to get past the compilation and submit the QASM to our API.

Proxies are now supported with the limitation that they will revert from websockets to polling of the restful interface. For long-running/long-queued jobs this may impact the response time somewhat, but it should be limited to a small fraction of the overall wait-time.

* Squashed commit including basis update and proxy support.

* Eliminated support for python 3.5, updated style to lint guidelines.

* Updates based on comments from PR on github.

* ReviewFix: Use logger for warning message.  Update comment that was incorrect.