Add missing super() to test setUp() methods
In recent qiskit-terra releases the base test class that terra exports
for some utility methods added a strict requirement that the base
class's setUp method is always called. This was done to ensure that some
fixture tracking around logging in parallel is always run. However, one
test class was not calling super().setUp() when it defined a setUp()
method which resulted in a test failure when running tests with the
latest terra (like we do in CI). This commit fixes this issue by adding
the missing super() call.