Workaround pylint limitations
In the recent terra 0.16.0 release the qiskit.providers module
docustring grew to include detailed documentation about the interface.
However, this change breaks the pylint parser (see
Qiskit/qiskit-ibmq-provider#784)  for more details and makes it unable
to do anything with module members off of qiskit.providers. To
workaround this bug in pylint this commit adds qiskit.namespace to the
ignored modules (just like those with c extensions) that pylint is
unable to handle correctly.