Add missing super() to test setUp() methods (#13)
* Add missing super() to test setUp() methods

In recent qiskit-terra releases the base test class that terra exports
for some utility methods added a strict requirement that the base
class's setUp method is always called. This was done to ensure that some
fixture tracking around logging in parallel is always run. However, one
test class was not calling super().setUp() when it defined a setUp()
method which resulted in a test failure when running tests with the
latest terra (like we do in CI). This commit fixes this issue by adding
the missing super() call.

* Workaround pylint limitations

In the recent terra 0.16.0 release the qiskit.providers module
docustring grew to include detailed documentation about the interface.
However, this change breaks the pylint parser (see
Qiskit/qiskit-ibmq-provider#784)  for more details and makes it unable
to do anything with module members off of qiskit.providers. To
workaround this bug in pylint this commit adds qiskit.namespace to the
ignored modules (just like those with c extensions) that pylint is
unable to handle correctly.