Update provider to versioned provider interface
This commit updates the honeywell provider to use the latest versioned
abstract interface from terra (which was introduced in 0.16.0). Terra
will be deprecating the legacy interface in the near future, so to get
in front of that we should update the provider to use the current
interface version. At the same time this fixes some compatibility issues
that arise when running with the latest version of terra (which is
necessary because the minimum version is bumped to ensure the new
provider interface exists).