Pivot CI to use github actions (#16)
* Pivot CI to use github actions

Previously we were using a combination of azure
pipelines (for windows testing) and travis (for all other
platform testing) to run ci. However recent changes
in travi's quota and billing for open source projects
means it's not a viable option anymore. This commit
migrates all the CI to be unified on github actions
which has sufficient quota to not have this problem.

* Remove travis and azure configs

* Fix tests

* Remove broken import