Fix handling of job and result metadata
In manually testing the job submission against the backend there were
several places where the job metadata was not getting updated correctly.
While this was not fatal and the job would execute successfully and the
counts objects were all correct, the metadata bout the job's result
would not properly set the config or header for a circuit based job.
This commit corrects this issue so that we're properly setting the
config and headers from the job config and circuit metadata.