Update provider to versioned provider interface (#14)
* Update provider to versioned provider interface

This commit updates the honeywell provider to use the latest versioned
abstract interface from terra (which was introduced in 0.16.0). Terra
will be deprecating the legacy interface in the near future, so to get
in front of that we should update the provider to use the current
interface version. At the same time this fixes some compatibility issues
that arise when running with the latest version of terra (which is
necessary because the minimum version is bumped to ensure the new
provider interface exists).

* Fix handling of job and result metadata

In manually testing the job submission against the backend there were
several places where the job metadata was not getting updated correctly.
While this was not fatal and the job would execute successfully and the
counts objects were all correct, the metadata bout the job's result
would not properly set the config or header for a circuit based job.
This commit corrects this issue so that we're properly setting the
config and headers from the job config and circuit metadata.