Pin importlib-metadata in CI (#22)
The recent release of importlib-metadata broke compatibility with
stevedore on Python 3.7. Stevedore is a dependency of a dependency
of stestr which we use for running tests in CI so this is blocking CI
from working and tests from running locally with Python 3.7. This
commit pins the importlib-metadata version in CI to unblock things.