Update CI configuration and supported python versions (#25)
This commit makes two primary changes. The first is to update the tox
usage to be compatible with tox>=4.0.0 and sets the minimum tox version
to be 4.4.0 to ensure users are on a compatible tox version. This should
fix CI failures we've been seeing since Decemember when tox 4.0.0 was
released and broke everything. The second change is to update the
supported python versions. The versions of python we support in this
package needs to be expanded to cover currently supported version by the
upstream Python community. Mainly 3.6 is EoL and this adds explicit
support for 3.10 and 3.11. I think after the next release of this
package we should also drop 3.7 support since it goes EoL this summer.